+++ driver-core/CHANGELOG.rst
@@ -5,6 +5,8 @@ CHANGELOG
 --------
 
 - [bug] Prevent race between cancellation and query completion (JAVA-614)
+- [bug] Prevent cancel and timeout from cancelling unrelated ResponseHandler if
+  streamId was already released and reused (JAVA-632).
 
 
 2.0.9: